requirements.txt
django
fastapi
matplotlib
farm-haystack
transformers
yake
